+++ adminforth/spa/src/components/MenuLink.vue
@@ -13,14 +13,28 @@
     <component v-if="item.icon" :is="getIcon(item.icon)" class="w-5 h-5 text-lightSidebarIcons dark:text-darkSidebarIcons  transition duration-75  group-hover:text-lightSidebarIconsHover dark:group-hover:text-darkSidebarIconsHover" ></component>
     <span class="ms-3">{{ item.label }}</span>
     <span v-if="item.badge" class="inline-flex items-center justify-center w-3 h-3 p-3 ms-3 text-sm font-medium rounded-full bg-lightAnnouncementBG dark:bg-darkAnnouncementBG 
-          fill-lightAnnouncementText dark:fill-darkAccent text-lightAnnouncementText dark:text-darkAccent">{{ item.badge }}</span>
+          fill-lightAnnouncementText dark:fill-darkAccent text-lightAnnouncementText dark:text-darkAccent"
+    >
+          
+        <Tooltip v-if="item.badgeTooltip">
+          {{ item.badge }}
+
+          <template #tooltip>
+            {{ item.badgeTooltip }}
+          </template>
+        </Tooltip>
+        <template v-else>
+          {{ item.badge }}
+        </template> 
+
+    </span>
 
   </RouterLink>
 </template>
 
 <script setup lang="ts">
 import { getIcon } from '@/utils';
-
+import { Tooltip } from '@/afcl';
 const props = defineProps(['item', 'isChild']);