+++ adminforth/spa/src/components/ResourceForm.vue
@@ -1,5 +1,7 @@
 <template>
   <div class="rounded-default">
+    <pre>
+  </pre>
     <form autocomplete="off" @submit.prevent>
       <div v-if="!groups || groups.length === 0">
         <GroupsTable
@@ -68,7 +70,6 @@ import { useI18n } from 'vue-i18n';
 
 
 const coreStore = useCoreStore();
-const { t } = useI18n();
 const router = useRouter();
 const route = useRoute();
 const props = defineProps({
@@ -80,7 +81,7 @@ const props = defineProps({
 
 const unmasked = ref({});
 
-const mode = computed(() => route.name === 'resource-create' ? t('create') : t('edit'));
+const mode = computed(() => route.name === 'resource-create' ? 'create' : 'edit');
 
 const emit = defineEmits(['update:record', 'update:isValid']);